+++ src/rocq_elpi_builtins.ml
@@ -1336,6 +1336,8 @@ end)
 module Summary = struct
   type 'a ref = 'a Stdlib.ref
   let ref ~name x = Summary.ref ~name x
+  let (!) = (!)
+  let (:=) = (:=)
 end
 [%%else]
 module Summary = struct