+++ Changelog.md
@@ -1,4 +1,9 @@
 
+# UNRELEASED
+
+### LIB:
+- New  `coq.indt-decl->ids`, `coq.mutual?`, `coq.mutual.members`
+
 # [3.5.0] 21/07/2026
 
 Requires Elpi 3.7.1 and Rocq 9.0, 9.1, 9.2 and 9.3.